refactor(bs): share refund ManualTransaction Prisma+Bulldozer dual-write
